% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{this is also just a fake description}

\item{n}{this is where we enter the number of values}

\item{p}{this is also just a fake description}
}
\value{
A straigh barplot
}
\description{
mybin
}
\examples{
mybin(iter=1000,n=10,p=0.7)
}
