% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{this is also just a fake description}

\item{x}{this is also just a fake description}

\item{fun}{this is also just a fake description}

\item{alpha}{this is also just a fake description}

\item{cx}{this is also just a fake description}

\item{...}{this is also just a fake description}
}
\value{
this is also just a fake description
}
\description{
myboot2
}
