% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Title}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{this is also just a fake description}

\item{iter}{this is also just a fake description}

\item{lambda}{this is also just a fake description}

\item{...}{This is a fake description}
}
\description{
Title
}
