% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Title}
\usage{
mymaxlik(lfun, x, param)
}
\arguments{
\item{lfun}{This is the function that you want to get the max likelihood of.}

\item{x}{This is your data that needs to be used.}

\item{param}{This is going to be a sequence and should be represented by possible values for your parameter.}
}
\description{
Title
}
