% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{this is also just a fake description}

\item{sigma}{this is also just a fake description}

\item{a}{this is also just a fake description}
}
\value{
graph with a curve
}
\description{
Title
}
