% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{'just a random plot function that is specific to lab 4'}
\usage{
myplot(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
numeric vector
}
\description{
'just a random plot function that is specific to lab 4'
}
