% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate Ticket Numbers}
\usage{
ntickets(N = 200, p = 0.95, gamma = 0.02)
}
\arguments{
\item{N}{The total number of tickets. Default is 200.}

\item{p}{The probability parameter. Default is 0.95.}

\item{gamma}{The gamma parameter. Default is 0.02.}
}
\value{
A list containing calculated ticket numbers and other related information.
}
\description{
This function calculates the minimum variance of ticket numbers
based on given parameters.
}
\examples{
ntickets()
}
