% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Title}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{This is a fake description}

\item{mn}{This is a fake description}

\item{...}{This is a fake description}
}
\value{
barplot with segments
}
\description{
Title
}
